linux_ionic: more code fixes seen while propping for upstream (#13273)
This is another set of little code tweaks that make the driver
more ready for upstream.

+++ platform/drivers/linux/eth/ionic/ionic_lif.h
@@ -161,7 +161,7 @@ struct lif {
 	unsigned int nrxq_descs;
 	u32 rx_copybreak;
 	unsigned int rx_mode;
-	u32 hw_features;
+	u64 hw_features;
 	bool mc_overflow;
 	unsigned int nmcast;
 	bool uc_overflow;
